netcfg: Rework removal of DNS resolver settings
When wanting to remove a DNS resolver settings object from the
NetCfg::DNS::SettingsManager, instead of triggering that via the
SettingsManager and remove the object instantly handle this inside the
ResolverSettings object itself.

Deleting the ResolverSettings object instantly removes the possibility
to catch these removals when it is needed to send NetworkChange signals
later on in the ApplySettings() method.  These signals should not be
sent before they have been applied to the system, so that these signals
are fairly within the same time window as when these changes happens
in reality.

Signed-off-by: David Sommerseth <davids@openvpn.net>