netcfg: Do not add duplicated DNS resolver settings
The Core library will in some reconnect situations re-add DNS settings
pushed from the server without initiating a flushing of prior settings.

This adds a check if the DNS name server or search domains have been
registered before by the ResolverSettings object and skip duplicated
entries.

Note that this is handled on a per ResolverSettings object.  If multiple
VPN sessions pushes overlapping information, it is up to the
ResolverBackendInterface implementation to evaluate what to do in these
situations.

Signed-off-by: David Sommerseth <davids@openvpn.net>