python: Extend ConfigParser to support embedding --auth-user-pass file
The --auth-user-pass option is a bit special.  It may be present as a
standalone option, or it can be given a filename containing username and
password to use for the user authentication.  In the latter case, we
will need to embedd this information into the configuration profile.

To handle this optional embedding, the ConfigParser.EmbedFile class
needed to be extended to support this use case.  And the profile
generator producing the merged configuration profile needed to learn how
to handle this use case as well.

Signed-off-by: David Sommerseth <davids@openvpn.net>