netcfgmgr/dns: Add indication when to apply DNS settings
Some DNS resolver settings backends needs to configure the DNS settings
after the tun interface has been configured.  The disadvantage with this
is that these settings may result in a short time window where DNS
queries may leak to DNS servers outside the DNS tunnel.  Due to this, we
enable the resolver backend to indicate when it wants to do the resolver
settings updates.

Signed-off-by: David Sommerseth <davids@openvpn.net>