build: Rework SELinux logic with --disable-selinux-build
This reworks the logic to not attempt doing any kinds of checks at
all if --disable-selinux-build is given to ./configure.

It also will explicitly check for /usr/share/selinux/devel/Makefile
in addition to ${datarootdir}.  Using ${datarootdir} is problematic
in the detection phase, as it is manipulated by --prefix.  But since
Fedora, RHEL and Debian SELinux development packages do ship the
required Makefile in /usr/share/selinux/devel/Makefile, this is a
safe extra check.

Signed-off-by: David Sommerseth <davids@openvpn.net>