ovpn3cli/session: Handle re-authentication for --resume/--restart
Before this change, the session-manage --resume and --restart
operations would just do the Resume() or Restart() calls to the
backend VPN session object and exit.  This would not work well
if the server requires a re-authentication.

This change refactors the logic for session-start and the
session-manage --resume and --restart operations to share the
same connection start logic to catch re-authentication scenarios
better.

Signed-off-by: David Sommerseth <davids@openvpn.net>