ovpn3cli/sessions: Improve SIGINT handling when starting/resuming sessions
Currently the SIGINT (CTRL-C) handling was fairly poor when a
user started, resumed or restarted a VPN session and hit CTRL-C.
What happened was that the openvpn3 program exited, but the
session manager never received any information that the user
aborted the operation and the backend VPN client process
continued to do what it was told to do.

This is particularly bad, as it could leave VPN sessions running
in the background.  If that backend process needed a
re-authentication, it would not do anything.  Or if a tunnel
was started by mistake and the server is not available, it would
try indefinitely to get a connection running.

This commit changes the current behaviour to respect a console
user sending CTRL-C during session-start, session-manage --resume
or --restart operations.  If the user does that before the
openvpn3 program has exited, the VPN session will be disconnected
and shutdown.

Signed-off-by: David Sommerseth <davids@openvpn.net>