sessionmgr: Implement sending SessionManagerEvent signals
This will make the Session Manager (net.openvpn.v3.sessions) broadcast
a SessionManagerEvent to all users on the system when a new OpenVPN 3
client session is created or removed/destroyed.  The signal contains
a type (SESS_CREATED/SESS_DESTROYED), a D-Bus path to the session
object and the UID of the owner of this session.

Signed-off-by: David Sommerseth <davids@openvpn.net>