client: Fix failing --log-file feature
The --log-file option in openvpn3-service-client would on some
distributions not work as intended.  The issue seems related to
newer compilers deleting the std::ostream object passed to
the StreamWriter object as a reference.  This was discovered
on Ubuntu 20.04.

To resolve this, make the std::ostream object a unique smart
pointer, which is sent to the StreamWriter object.  We cannot
initialize the std::ostream object outside the if() block
related to --log-file, as it depends on the contents of the
--log-file argument.

Signed-off-by: David Sommerseth <davids@openvpn.net>