netcfg: Avoid segfault when configured without --resolv-conf
Before this change, openvpn3-service-netcfg would crash when a VPN
session was started and the netcfg service was not configured to
use --resolv-conf.

This was due to a missing check when instantiating the NetCfgDevice
object, where it tried to access a resolver object which was
pointing at nullptr.

Extended with additional checks, to also ensure accesses to
invalid dnsconfig objects would also be considered.

Signed-off-by: David Sommerseth <davids@openvpn.net>