dbus/proxy: Migrate to better value extraction in GetProperty() and CheckServiceAvail()
The g_variant_get() may with some compilers make the code fragile and
unreliable and hard to debug with compiler optimization reduced.  This
feels like a Glib2 issue, as using a different approach to extract the
values does not provide the same kind of issues.

This has been done a few times earlier already, but this time it was
these two methods making debugging life harder.

Signed-off-by: David Sommerseth <davids@openvpn.net>