build: Clean up after version metadata rework
In commit 769bbdc34bac39e55a the config-version.h handling was reworked
and now only utils.cpp depends on it.  Still a few targets had a
reference to this version header in Makefile.am, so clean up that.

In addition add config-version.h to .PHONY.  This will recreate
config-version.h on each build.  This will cause utils.o to be
recompiled and all binaries relinked.  This is acceptable, as utils.cpp
is compiled quickly and the linking process goes fast.  The benefit is
that the version information in all programs contains up-to-date version
information.

Signed-off-by: David Sommerseth <davids@openvpn.net>