client: Process CLIENT_HALT event
The OpenVPN server can in some cases forcefully disconnect an already
connected client.  This triggers the CLIENT_HALT event, which was not
processed before.

Without this change, the openvpn3-service-client process will just
shutdown quite brutally via the LogFATAL() call if the ev.fatal flag is
true.  This is set to true with CLIENT_HALT.

Instead we will just handle CLIENT_HALT as a basic failed connection.
Since it was not triggered by the user the StatusMinor::CONN_DISCONNECTED
is not appropriate.

Signed-off-by: David Sommerseth <davids@openvpn.net>