client: Fix incorrect environment array handling
In commit 592403d13, openvpn3-service-backendstart was extended to pass
environment variables to the child process it execve()'s.  The handling
of empty environment variables was incorrect and didn't work as
intended.

Instead use a different approach to ensure that the environment variable
pointer is really nullptr if no environment variables has been
configured.

Signed-off-by: David Sommerseth <davids@openvpn.net>