cli/netcfg: Make netcfg-service command more robust
The 'openvpn3-admin netcfg-service' command would just exit with a
verbose error message if it could not connect to the
net.openvpn.v3.netcfg service, before it would do anything at all.  This
does not work well with --config-file-override where the intention is to
be able to modify the configuration file regardless if the netcfg
service can start or not.

Signed-off-by: David Sommerseth <davids@openvpn.net>