common: Extend Configuration::File parser with option groups
This is similar to the ParsedArgs::CheckExclusiveOptions(), but the
groups are declared the Configuration::OptionMapEntry() directly.

The Configuration::File class has been extended with its own
CheckExclusiveOptions() method which will throw an ExclusiveOptionError
exception if more options belonging to the same exclusive group are used
in the configuration file.

This check needs to be called explicitly in the code implementing
Configuration::File.

Signed-off-by: David Sommerseth <davids@openvpn.net>