netcfg: systemd-resolved - Make all global queries also go via VPN
Without this setting, global DNS queries would not go via the VPN
connection; only domains where a search domain would be pushed by the
server.

This will need to be improved and made configurable per VPN profile
basis, as well as controllable via --push options from the server.  But
this behaviour is closer to the current default behaviour in OpenVPN.

Signed-off-by: David Sommerseth <davids@openvpn.net>