netcfg/resolved: Better exception handling when communicating with systemd-resolved
When the proxy communication with the systemd-resolved service failed,
various exceptions where not handled properly.  Sometimes we can accept
that a property is not available, and otherwise use more specific
exceptions.

This also extends the netcfg-systemd-resolved-basic test program to be
able to extract and display as much information as possible.  Older
systemd-resolved versions might not have all properties or methods
available as we expect.

Signed-off-by: David Sommerseth <davids@openvpn.net>