python/shell: Add bash-completion for openvpn2
This extends the openvpn3.ConfigParser class to gather information
useful for bash-completion scripts.  And it adds a generator script
which will generate an openvpn2 bash-completion script containing
all the options openvpn2 supports, based on the options/arguments
added to the ConfigParser class.

The reason for creating a static openvpn2 script was to avoid instead
of using the dynamic approach used by openvpn3, was simply to avoid
kicking off a Python process each time bash-completions gets called
for the openvpn2 command.

Signed-off-by: David Sommerseth <davids@openvpn.net>