common: Fix Configuration::File::Save() not writing empty files
A check had been added to the Save() method to just return if it
would write an empty configuration file.  This made it impossible
to remove all set values and update the configuration file.

This fixes an issue with openvpn3-admin netcfg-service when
all the entries in a config file is removed by using --config-unset.

Signed-off-by: David Sommerseth <davids@openvpn.net>