cli/admin: netcfg-service could exit with an empty error
If the net.openvpn.v3.netcfg service was not running and it would
start a bit slow on a host, the 'openvpn3-admin netcfg-service' command
would often just exit with "** ERROR ** ():".

The issue was that the NetCfgProxy::Manager() object did not call the
CheckServiceAvail() method when establishing the proxy object.  This
call will ensure net.openvpn.v3.netcfg is available before continuing.

Signed-off-by: David Sommerseth <davids@openvpn.net>