configmgr: Better handling of incorrect configuration profiles
When importing configuration files via D-Bus calls or during startup,
the openvpn3-service-configmgr could SIGABRT if the configuration profile
was invalid or corrupt.

This adds an improved exception handling on a few critical steps during
the configuration import and logs or reports an error if the configuration
profile could not be imported.

Signed-off-by: David Sommerseth <davids@openvpn.net>