configmgr: Add DCO device naming hack
Currently the Core library does not provide a feature to set the
interface name of the ovpn-dco devices.  It defaults to the value
provided in --dev in the configuration profile, which is typically
named 'tun'.

This is just a temporary hack to better identify virtual interfaces
using the ovpn-dco interface on the system.  So if the DCO feature
has been enabled on a configuration profile, the configuration
returned by the Fetch() method will have a slight tweak replacing
the --dev option and adding --dev-type if not set.

Signed-off-by: David Sommerseth <davids@openvpn.net>