client: Properly reset the paused flag on session restart
If a paused session was restarted instead of resumed, the pause
flag was not reset.  This led to confusing error messages and
inabilities to pause the session.  And resuming a restarted
session also leads to other issues.

Since it's considered fine to restart a paused session, the
fix is really simple and obvious.

Signed-off-by: David Sommerseth <davids@openvpn.net>