python: Resolve error in openvpn2 on disconnect with pre-closed sessions
When openvpn2 is running with a session in the foreground, it listens
for StatusChange and Log events from the backend, displaying them on
the console as they appear.  When CTRL-C is hit, it retrieves
statistics and disconnects.

If the same session has been closed via other approaches (for example
openvpn3 session-manage --disconnect), a D-Bus exception would be
thrown when retrieving the statistics or calling the disconnect
method.

This is fixed by catching the '[...].DBus.Error.UnknownMethod'
exception and ignoring it.  All other exceptions are raised as
proper exceptions.

Signed-off-by: David Sommerseth <davids@openvpn.net>