python: Add --mute, --route-delay and --route-method to ignored options
These options are not used nor supported by OpenVPN 3 Core library nor
OpenVPN 3 Linux.  The logging is handled very differently from
OpenVPN 2.x, so the --mute option does not serve any purpose.

The --route-delay and --route-method options are typically used on
Windows to switch how routes are configured and delay the setup so the
TUN/TAP interface has managed to be initialized first.  This is not
needed on OpenVPN 3 Linux, as the openvpn3-service-netcfg service takes
care of all of this and doing things in the right order and time.

Since all of these three options does not break any configuration, they
will just be ignored instead of being handled as an error.

Signed-off-by: Romain Loutrel <un.romain.loutrel@gmail.com>
Signed-off-by: David Sommerseth <davids@openvpn.net>