netcfg: Implement DNS resolver scope support
This implements the needed configuration settings to properly set up the
global or tunnel DNS resolver scopes in systemd-resolved.  For the
resolv-conf based mode, this setting will be ignored (and logged as a
warning) as resolv-conf only supports a single DNS resolver.

Currently, the tunnel scope is very simplistic and will only result in
systemd-resolved not querying anything than pushed DOMAIN/DOMAIN-SEARCH
domains via the DNS servers for the VPN session.  It is more a reverse
filter, than an explicit filter.  All other "global" scopes will also be
queried.  Domains *not* matching the DOMAIN/DOMAIN-SEARCH settings
will be *not* queried via this VPN link.

This behaviour is more how systemd-resolved is designed and this
behaviour will hopefully be further improved later on.

Signed-off-by: David Sommerseth <davids@openvpn.net>