tests/signal_listener: Add support for Log() with session tokens
The Log() D-Bus signal from openvpn3-service-client may often also
carry a session token in the Log event, this was not supported in the
signal_listener debugging utility.

Rework the logic a bit to detect the format before trying to parse the
signal and avoid bailing out on incompatible signals types.

Signed-off-by: David Sommerseth <davids@openvpn.net>