python: Extend ConfigParser with DCO support
This adds --disable-dco, which is expected to be a new option in
OpenVPN 2.6, to disable usage of the DCO interface.

Since OpenVPN 3 currently does not automatically detect the DCO
capabilities of the host, also add --enable-dco to stay cohorent with
the option syntax of disabling DCO.

Both of these options are also added to a new option group,
tech-preview, to highlight that these options are not production ready
and may change.

Signed-off-by: David Sommerseth <davids@openvpn.net>