core-ext: Fix improper processing of potential inline options
Some options may be used with different syntaxes, like --auth-user-pass.
This can be used as a single stand-alone option, with a file argument or
as a inline option (<auth-user-pass>...</auth-user-pass>).

Before this change, it would expect it to only be used as a inline
option.  This changes the OptionListJSON and ProfileMergeJSON Core
extensions to account for the various syntaxes.

The test programs for converting to/from JSON has also been slightly
improved to use also use the ProfileMerge implementation to ensure test
configurations referencing a file is handled correctly as well.

Signed-off-by: David Sommerseth <davids@openvpn.net>