selinux: Rework the policy build Makefile
The current approach of including the main SELinux devel Makefile had
several issues and when adding a new policy into the mix, it got more
complicated - especially with out-of-tree builds.

Redoing the policy building to call 'make' directly using the SELinux
Makefile. This seems to avoid these issues and is more in line with how
policy building is done elsewhere.

Also reworked and made the Makefile.am more generic.  Adding a new
policy is just a matter of adding a new new .pp file reference.

Signed-off-by: David Sommerseth <davids@openvpn.net>