netcfg/resolved: Check if org.freedesktop.PolicyKit1 is available
The systemd-resolved integration will not work unless polkit is
available.  Since the NetCfg service runs as an unprivileged user, the
policy we provide (src/policy/net.openvpn.v3.rules) allows the openvpn
user to call certain methods in the org.freedesktop.resolve1 service.

If the polkit service is not available, NetCfg will be denied access to
the org.freedesktop.resolve1.  Thus the DNS configuration will fail with
"Access Denied" errors.  So we try to catch that as early as possible
and ensure this is picked up and reported.

Signed-off-by: David Sommerseth <davids@openvpn.net>