sessionmgr: Split out pure manager functions from OpenVPN3SessionProxy
This moves the pure session manager functions, not operating on a single
session into its own main class.  A few of the methods here has been
rewritten to return a pointer to an OpenVPN3SessionProxy object
for a specific D-Bus object for the session, or an array/vector of such
pointers.

Signed-off-by: David Sommerseth <davids@openvpn.net>