cli/sessionmgr: Simplify property extraction
Since the sessionmgr service now will properly return access denied on
all properties users not granted access, the whole property retrieval
can build on the generic access denied logic.

Also add a simple is-this-run-by-root check, to further reduce code
complexity in the for() loop to not end up in a ugly odd-looking
listing.

Signed-off-by: David Sommerseth <davids@openvpn.net>