netcfg/resolved: Don't configure --dhcp-option DOMAIN as routing domains
When configuring search domains as routing domains in systemd-resolved,
it does not work as expected for many users.  It ensures that DNS
queries are sent only to the VPN provided DNS servers.  But most users
expects it to behave like traditional search domains.

Until we get a better specification on DNS resolver configuration in the
OpenVPN documentation, we accept that this is a better user experience
for the time being.

Signed-off-by: David Sommerseth <davids@openvpn.net>