cli/session: Not all connection failures are timeout related
Commit 6780ac3fec4b6afdd changed the behaviour to not have a timeout by
default.  But when adding --timeout it would behave like before with a
user set timeout.

This gave the side effect of reporting more connections issues as
timeout related while not being the case.  This improves the error
handling further to ensure only timeout related issues are reported with
a timeout reference.

Signed-off-by: David Sommerseth <davids@openvpn.net>