log: Improve LogEvent formatting
Make the GVariant* object generation more clever by considering if a
particular format should be used.  The default behaviour is to include
the session token value if present.  Otherwise it will be ignored.  This
can be overriden by setting the 'format' member accordingly.

This also extends the GVariant* parsing to consider the format received.

Signed-off-by: David Sommerseth <davids@openvpn.net>