client: Don't switch to Reconnecting state on initial connect
If the client could not reach a connection instantly, the OpenVPN 3 Core
client would try again to connect again and would send the
"RECONNECTING" event in this case.  From openvpn3-linux perspective,
this is not a reconnection.  So we don't change the session status from
CONNECTING and also does not send the reconnecting log event.

Once the client has reached a connected state, then it flips back to
reporting reconnecting statuses again.

Signed-off-by: David Sommerseth <davids@openvpn.net>