sessionmgr: Fix incorrect LogEvent proxy format
In commit f1c6b3038b16e the LogEvent formatting logic was modified to
consider the LogEvent::format setting.  This had the side-effect log
messages being proxied to the front-end applications to have the wrong
signature, and the Python based front-ends would stop working.

The fix is simple, set the approriate format setting when wiping the
session token from the proxied log event.

Signed-off-by: David Sommerseth <davids@openvpn.net>