configmgr: Extend with session ownership transfer flag
A configuration profile where the owner has granted access to the
profile to another user, that user can start a new VPN session.  When
this flag is not set, the user starting the session will own the VPN
session and manage completely on its own.

By setting this flag to true, the ownership of the VPN session is
transferred to the same owner as the configuration profile and will just
grant a limited possibility to manage the session by the user starting
the session.

Signed-off-by: David Sommerseth <davids@openvpn.net>