sessionmgr: Respect the configuration profile transfer-ownership flag
This implements transferring the ownership of the session to the same
user owning the VPN configuration profile, if enabled in the profile.

To ensure the user starting the VPN session is able to fully connect and
interact with the session during the session startup, the initial user
is a granted basic session management ACL and log access.

Signed-off-by: David Sommerseth <davids@openvpn.net>