sessionmgr: Retrieve the client log-level for the session log-level
With recent changes, the VPN client backend log-level can now be
adjusted via --verb in the configuration file or the log-level override
in the configuration profile.  But this was not reflected in the session
object itself.

Retrieve the log-level the client backend uses for the session and set
that value in the session object.

Signed-off-by: David Sommerseth <davids@openvpn.net>