core-ext: Fix incorrect handling of --static-challenge in JSON export
The Option::render() method in the OpenVPN 3 Core library does not
produce a correct rendering of the --static-challenge option.  This
option takes to arguments, the static challenge text displayed to the
user - and this must be encapsulated with double quotes if the challenge
contains spaces.  And at the end there is an echo flag, which can be 0
or 1.

The Option::render() method does not add the needed double quote
encapsulation on the challenge text.  The alternative is
Option::escape() which does add the double quotes, but it also puts the
echo flag inside the challenge text.

For now we just add a quick hack, where the echo flag is extracted and
the challenge text is properly encapsulated with double quotes for the
static-challenge option alone.

Signed-off-by: David Sommerseth <davids@openvpn.net>