common/shell: Fix bash-completion for options with optional arguments
Options with optional arguments need to have a '=' to provide a value to
the option.  The shell-completion logic and bash-completion parsing had
needed to be extended to better support this use case.

This also removes value suggestions to options with optional arguments
when the '=' is not present.  Value suggestions comes only when the '='
separator is present.

Signed-off-by: David Sommerseth <davids@openvpn.net>