cli/session: Extend the session module with session-auth
This new command will be used to list and interact with on-going VPN
sessions requiring user interaction to complete authentication.

The use case is for sessions needing re-authentication after the first
connection succeeded or that the session was started in the background
and the server required the user to provide more credentials.

The current implementation only lists a primitive list of sessions
requiring user interaction.  This will be extended by the following
commits.

Signed-off-by: David Sommerseth <davids@openvpn.net>