client: Change default log-level to 3
Since the openvpn3-service-client process will now parse both the --verb
option in configuration files, it can be changed by setting the
'log-level' override in the configuration profile and it can be adjusted
on a per-running-session via 'openvpn3 session-manage --log-level' there
should be enough approaches to get a reasonable log verbosity.  No need
to send more details before it's needed.

Signed-off-by: David Sommerseth <davids@openvpn.net>