python: Extend openvpn3-as with --owner
This option is available when run as root, where it allows the
downloaded configuration profile to have a different owner than the root
user (default behaviour).  This is restricted to root, as root is the
only user allowed to transfer ownership.

This is in particular useful when used together with --systemd-start,
which grants the owner of the configuration profile to manage the VPN
session started at boot.  This is how the VPN session owner is
transferred when root starts the VPN session.

Signed-off-by: David Sommerseth <davids@openvpn.net>