client: Extend RegistrationConfirmation D-Bus method with session path
Before this change, the client process would always report just a static
object path which does not give much useful information in the signals
sent via the BackendSignals object.

This changes the RegistrationConfirmation D-Bus method, which the session
manager calls, to provide the real session path used by the session
manager.  This path is then set in the BackendSignals object and allows
all the next signals emitted to have a clear reference to the session in
the session manager.

This change also replaces the GLib2 g_variant_get() call with our own
GLibUtils approach extracting the parameters in
RegistrationConfirmation.

Signed-off-by: David Sommerseth <davids@openvpn.net>