log: Extend D-Bus logging with path filtering
Extend the LogFilter API with methods to set path filters to be used
when proxying log and status events.

The LogSender:ProxyLog() method is extended with an optional path
variable, which is expected to contain the path of the sending D-Bus
object.  This path will be processed via the LogFilter path filtering,
for decision if the message should be forwarded or not.

Signed-off-by: David Sommerseth <davids@openvpn.net>