sessionmgr: Re-implement log forwarding to end-users
In a prior commit, the log forwarding was teared out to make it easier
to implement a better solution which passes some the responsibility to
forward log events to end-users from the session manager to the logger
service.

This implements a new net.openvpn.v3.sessions.LogForward method which an
end-user program can call to receive log and status change events.  This
implementation should also respect the 'allow-log-access' setting, which
the prior implementation did not fully do.

Signed-off-by: David Sommerseth <davids@openvpn.net>