ovpn3cli: Re-implement log command with new logging infrastructure
The previous log fetching implementation has been completely redone and
is no longer proxied via the session manager, the 'openvpn3 log' command
needed to be adopted accordingly.

This patch refactors out the LoggerBase and adopts it to the plain
Logger class which is now only used for the --config-events mode.

For session logging, the new LogForwardBase implementation is used.  And
since this does no longer require the session proxy for receiving log
events, the create_session_proxy() has been renamed and simplified -
including no longer needing the cleanup() method.

Signed-off-by: David Sommerseth <davids@openvpn.net>