python: Rework openvpn3.SessionManager for the new logging infrastructure
With the recent changes to the logging infrastructure, the Log and
StatusChange event signals needs to be handled differently as they are
now issued by the net.openvpn.v3.log service.

This new approach also requires the log forwarding to be disabled before
closing down.  Add an additional destructor to the session object to
handle this properly.

Signed-off-by: David Sommerseth <davids@openvpn.net>