client: Rework BackendSignals to avoid Log signal duplication
Before the changes to the net.openvpn.v3.log service, log events would
also be sent to the session manager by default - in addition to the log
service.  This behaviour is no longer needed, as the logger service will
take care of further proxying Log and StatusChange events to other
listeners.

This change removes this signal duplication for Log events.  Only
StatusChange signals are sent to both the logger and sessionmgr services
now.  The sessionmgr service depends on these signals to know the state
of the VPN session.

Signed-off-by: David Sommerseth <davids@openvpn.net>