shell: Fix trailing spaces in bash-completion
When tab-completing directories, an additional space was added after the
directory name.  This should not happen, but is a result of how the
'complete' setup in bash works.

Instead we now tell bash to not add any spaces afterwards.  That has some
other behaviour changes, resulting in needing more often a double-tab to
move further.  For already completed commands, we can in most cases
start with a leading double-dash, to reduce the need for extra tabs in
some places.

Signed-off-by: David Sommerseth <davids@openvpn.net>