log: Fix memory corruption with syslog/openlog()
The openlog() takes a char* as a program identifier.  This pointer is
not preserved within the syslog calls.  Since the SyslogWriter only
passed on the pointer, if the pointer outside of SyslogWriter changed it
could lead to pointing to an invalid address.

This refactors the SyslogWriter() class to be based in std::string where
it will keep a local char* buffer containing the identifier.

Signed-off-by: David Sommerseth <davids@openvpn.net>