cli/log: Add new options for logger config file management
So far openvpn3-service-logger has only had some runtime settings which
would be preserved if the --state-dir option was present.

With this change, the openvpn3-admin log-service command has been
extended to also be able to manage other openvpn3-service-logger
settings which requires the process to restart before becoming active.

Signed-off-by: David Sommerseth <davids@openvpn.net>