client: Send platform OS/distro peer information to server
This adds sending the IV_PLAT_VER peer-info field to the VPN server when
the --push-peer-info option is used.  This will contain the string
returned by the PlatformInfo class.

For this to work, the SELinux policy needed to be extended, to allow the
openvpn3_client_t labelled process to communicate with the
systemd_hostnamed_t label; the PlatformInfo class attempts to query the
org.freedesktop.hostname1 service for distro details.

Signed-off-by: David Sommerseth <davids@openvpn.net>