log: Pass the LogTag objects as smart pointers
The previous implementation had a hacky constructor in the
LogMetaDataValue::create() method for string values, where an empty
LogTag object was created in the LogMetaDataValue constructor.  The
clang++ compiler complained (rightfully) about potential issues with
this.

This wouldn't necessarily pose a problem in the code, as we depend on
the LogMetaDataValue::Type whether to process the value as a LogTag
or std::string.  But passing a Ptr to the LogTag seems cleaner and more
correct.

Signed-off-by: David Sommerseth <davids@openvpn.net>