client: Add workaround for OpenVPN Access Server web authentication
When the OpenVPN Access Server is configured to do web based
authentication, the OpenVPN 3 Core library is not given a username by
the user itself, it is extracted from the configuration profile as "meta
options".  Such options are found in the comment lines and in the Access
Server context, it is prefixed with "OVPN_ACCESS_SERVER".

OpenVPN 3 Linux did not consider these meta options at all.  But it
needs to since the OpenVPN 3 Core library handles the server pushed
--auth-token differently than OpenVPN 2.x; it will not send the token
value during renegotiations unless a username is set.

This will be improved in a later release of OpenVPN 3 Core library, but
for now we add this hack to work around this issue in OpenVPN 3 Linux.

Signed-off-by: David Sommerseth <davids@openvpn.net>