log: Fix missing O3_LOGTAG meta data variable
The O3_LOGTAG should always be added to the journald logging. That was
not happening, as the LogWriter::AddMeta() considers if the meta logging
has been enabled or not.  This happens before the log event is processed
by the LogWriter implementation, which can also check if the meta
logging is enabled.

Since the LogWriter::AddMeta() implementation for LogTag objects was
only used on place, simplify this code and specialise a new method for
LogTag objects.  This ensures this will always be included, as long as
the LogWriter implementation in use (SyslogWriter, StreamWriter,
JournaldWriter) loops through the meta data variables.  For the
JounraldWriter, this will always happen.

Signed-off-by: David Sommerseth <davids@openvpn.net>