log: Extend LogEvent with str() method
The new str() method can be used to retrieve the LogEvent as a formatted
string, extending the possibilities so far only available the
operator<<() method.

A new LogEvent member has been added, indent_nl.  This is used to indent
log lines before new lines in the input log message.  The width of the
indenting is controlled by this variable.  The default is 0.

Two str() methods are added.  One takes no default values and returns the
same formatting the operator<<() method used to result in, if indent_nl
is set to 0.  Otherwise, it will respect the indent_nl setting.

The second str() method takes one or two arguments.  The first argument
is the indent width, which overrides the indent_nl setting.  The second
argument is a boolean which can be used to disable the LogPrefix() call
which includes the LogGroup/LogCategory strings.  If the last argument
is not provided, it defaults prefix the log message with LogPrefix().

Signed-off-by: David Sommerseth <davids@openvpn.net>