ovpn3cli: New command - openvpn3-admin journal
This command requires to be run as root, as it needs to be
able to access the systemd-journal log entries produced by the
OpenVPN 3 Linux stack.

Running it will retrieve only the OpenVPN 3 Linux log entries
stored by the systemd-journald service.  It will require the
openvpn3-service-logger service to be configured to use the
journald logging facility.

Signed-off-by: David Sommerseth <davids@openvpn.net>